test(ui): split custom theme ownership coverage
